+++ src/FabPermissions.php
@@ -52,7 +52,7 @@ class FabPermissions extends Plugin
      *
      * @var string
      */
-    public $schemaVersion = '1.0.0';
+    public $schemaVersion = '1.0.1';
 
     // Public Methods
     // =========================================================================