+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 1.0.4 - 2019-06-18
+### Fixed
+- Fixed a bug where admin permissions weren't respected when no user groups existed
+
+### Modified
+- Updated Fab service to handle admin permissions
+- Updated migration to let userGroupId be NULL
+- Updated JS to set admin permissions
+
 ## 1.0.3 - 2019-06-18
 ### Modified
 - Fixed LICENSE.md